Fix some codeclimate complaints
